requirements.txt
PyQt6>=6.4.0
pygame>=2.1.2
yt-dlp>=2023.3.4
python-distro>=1.8.0
pyaudio>=0.2.13; platform_system != "Linux" or python_version < "3.11"
pip>=23.0.0
setuptools>=65.5.1
wheel>=0.38.0 